refacture image_srcset method to be reusable in plone.namedfile
